feat: display task list for authenticated users only
